Add visitor to the program
Build the antlr4 support files using --visitor flag and
include the visitor into the C++ main program, call it
to show that visitor parses tree. This used a version of the
visitor that just printed out each node that it visited: this
would be replaced by inheriting from the visitor and making
our own visitor so the antlr4 generated code was not editied.